Add files via upload
This Python program is like a secret message tool! It lets you either turn a message into a secret code or turn a secret code back into a message. It's kind of like a secret decoder ring from spy movies.

Here's how it works:

When you choose to "Encrypt", you type in a message you want to turn into a secret code and how much you want to shift each letter in the message. For example, if you want to turn "hello" into a secret code with a shift of 1, it becomes "ifmmp".

When you choose to "Decrypt", you type in the secret code and how much it was shifted. It turns the secret code back into the original message. So, if you type in "ifmmp" with a shift of 1, it becomes "hello" again.

You just follow the instructions on the screen, and it helps you make secret messages and decode them. It's like being a spy with your own secret language!